refactor(native): split runtime camera modules into subdomain folder
