+++ src/native/engine/engine.c
@@ -12,8 +12,8 @@
 #include "runtime/engine/engine_lifecycle.h"
 #include "runtime/engine/engine_tick.h"
 #include "runtime/engine/gameplay_service.h"
-#include "runtime/camera_follow.h"
-#include "runtime/camera_basis.h"
+#include "runtime/camera/camera_follow.h"
+#include "runtime/camera/camera_basis.h"
 #include "runtime/input/move_target_service.h"
 #include "runtime/player/player_runtime_service.h"
 #include "runtime/player/player_gateway_abi.h"
