+++ src/native/engine/runtime/camera/camera_follow_policy.h
