+++ src/native/engine/runtime/player/player_runtime_service.c
@@ -1,6 +1,6 @@
 #include "player_runtime_service.h"
 
-#include "../camera_follow.h"
+#include "../camera/camera_follow.h"
 #include "../input/input_contract.h"
 #include "player_motion_host.h"
